chore: refactor some code
